backend/requirements_cloud.txt
# --- Core Web Server ---
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0

# --- Data & Excel ---
pandas
openpyxl
numpy

# --- Text & Image ---
markdown
pillow

# --- Other Dependencies (Kept from your list) ---
contourpy
cycler
fonttools
joblib
kiwisolver
matplotlib
packaging
pyparsing
python-dateutil
pytz
scikit-learn
scipy
six
threadpoolctl
tzdata
fake-useragent
tenacity
undetected-chromedriver
beautifulsoup4
httpx
gunicorn